requirements.txt
flask>=3.0
python-docx>=1.1
lxml>=5.0
